chore(liveops): ghi gói G-I18N (đợt W3.5) vào bảng quyền ghi, check-ownership đọc được đợt lẻ
Áp quyết định user 16/9/2026 vào phần máy đọc của kế hoạch:

- ownership.tsv: thêm gói G-I18N (W3.5, cỡ L, phụ thuộc toàn bộ W0–W3), 22 dòng
  file mới (Editor/Hub/Foundation/Language/*, Shell/LiveOpsHubLanguageMenu.cs,
  Tests/.../Foundation/Language/*), 2 dòng thư mục mới, 25 dòng cấp quyền ghi cho
  G-I18N trên file đang có (11 file LiveOpsHubStrings.*.cs, InterimUnavailableHubActions.cs,
  header/window/uxml/uss/classnames của Shell, 4 file capture, 4 file tools).
- check-ownership.py: số đợt đọc bằng float thay int — "W3.5" làm int() ném.
  --wave-deps vẫn báo đúng 4 lỗi có chủ đích (W1×2, W2×2), không thêm lỗi mới.

Kế hoạch (ngoài repo) đã sửa cùng nhịp: 0.3 PD-45, 1.1, 2.3, 2.4, 3.3, 8.10, 9.2,
9.5, 10.2, 10.3, 10.4, 13, 14 (V-28) + quyết định SPIKE-B dự phòng vào 7.0, 8.6,
8.7, 4.3, 11.2. Đặc tả gói: PLAN/i18n/G-I18N-SPEC.md.

+++ tools/liveops-hub/check-ownership.py
@@ -207,7 +207,8 @@ def check_wave_dependencies(table):
     errors = []
     for package in table.package_order:
         wave = table.packages[package]["wave"]
-        wave_number = int(wave[1:])
+        # Đợt có thể lẻ (W3.5 = gói G-I18N chen giữa W3 và W4) nên đọc bằng float, không int.
+        wave_number = float(wave[1:])
         for dependency in table.packages[package]["dependencies"]:
             if re.match(r"^W\d+$", dependency):
                 dependency_wave = dependency
@@ -216,7 +217,7 @@ def check_wave_dependencies(table):
                 if dependency_wave is None:
                     errors.append("%s phụ thuộc '%s' không có trong bảng" % (package, dependency))
                     continue
-            if int(dependency_wave[1:]) >= wave_number:
+            if float(dependency_wave[1:]) >= wave_number:
                 errors.append("%s (%s) phụ thuộc %s (%s) — phải ở đợt trước" % (package, wave, dependency, dependency_wave))
     if errors:
         for error in errors:
